♻️ Refactor EnhancedReactTable * Use composition instead of creating a fork of the react-table checkbox HOC
